+++ resources/js/views/AdminServerPoolsView.vue
@@ -267,7 +267,9 @@ const serversMultiselectRef = ref(false);
  */
 onMounted(() => {
   load();
-  loadServers();
+  if (!props.viewOnly) {
+    loadServers();
+  }
 });
 
 /**