+++ tests/Frontend/e2e/RoomsViewDescription.cy.js
@@ -1,5 +1,8 @@
 import { interceptIndefinitely } from "../support/utils/interceptIndefinitely.js";
-
+import {
+  clearTiptapContent,
+  selectTiptapContent,
+} from "../support/utils/tiptapHelper.js";
 describe("Rooms view description", function () {
   beforeEach(function () {
     cy.init();
@@ -28,7 +31,7 @@ describe("Rooms view description", function () {
     cy.get('[data-test="room-description-viewer"]')
       .should("be.visible")
       .within(() => {
-        cy.get("p").contains("Room description").should("be.visible");
+        cy.get("p").should("have.text", "Room description").and("be.visible");
       });
 
     cy.get('[data-test="room-description-edit-button"]').should("not.exist");
@@ -59,7 +62,7 @@ describe("Rooms view description", function () {
     cy.get('[data-test="room-description-viewer"]')
       .should("be.visible")
       .within(() => {
-        cy.get("p").contains("Room description").should("be.visible");
+        cy.get("p").should("have.text", "Room description").and("be.visible");
       });
 
     cy.get('[data-test="room-description-edit-button"]').should("not.exist");
@@ -87,7 +90,7 @@ describe("Rooms view description", function () {
     cy.get('[data-test="room-description-viewer"]')
       .should("be.visible")
       .within(() => {
-        cy.get("p").contains("Room description").should("be.visible");
+        cy.get("p").should("have.text", "Room description").and("be.visible");
       });
 
     cy.get('[data-test="room-description-edit-button"]')
@@ -120,7 +123,7 @@ describe("Rooms view description", function () {
     cy.get('[data-test="room-description-viewer"]')
       .should("be.visible")
       .within(() => {
-        cy.get("p").contains("Room description").should("be.visible");
+        cy.get("p").should("have.text", "Room description").and("be.visible");
       });
 
     cy.get('[data-test="room-description-edit-button"]')
@@ -145,7 +148,7 @@ describe("Rooms view description", function () {
     cy.get('[data-test="room-description-viewer"]')
       .should("be.visible")
       .within(() => {
-        cy.get("p").contains("Room description").should("be.visible");
+        cy.get("p").should("have.text", "Room description").and("be.visible");
       });
 
     cy.get('[data-test="tip-tap-editor"]').should("not.exist");
@@ -157,10 +160,12 @@ describe("Rooms view description", function () {
     // Check that editor shows the correct description
     cy.get(".tiptap")
       .should("be.visible")
+      .and("have.focus")
       .within(() => {
-        cy.get("p").contains("Room description").should("be.visible");
+        cy.get("p").should("have.text", "Room description").and("be.visible");
       });
-    cy.get(".tiptap").clear();
+
+    clearTiptapContent();
 
     // Cancel editing
     cy.get('[data-test="room-description-cancel-edit-button"]')
@@ -172,23 +177,25 @@ describe("Rooms view description", function () {
     cy.get('[data-test="room-description-viewer"]')
       .should("be.visible")
       .within(() => {
-        cy.get("p").contains("Room description").should("be.visible");
+        cy.get("p").should("have.text", "Room description").and("be.visible");
       });
     cy.get('[data-test="tip-tap-editor"]').should("not.exist");
 
     // Open editor again and check that description stayed the same
     cy.get('[data-test="room-description-edit-button"]').click();
+    cy.get('[data-test="room-description-edit-button"]').should("not.exist");
 
     cy.get('[data-test="tip-tap-editor"]').should("be.visible");
 
     cy.get(".tiptap")
       .should("be.visible")
+      .and("have.focus")
       .within(() => {
-        cy.get("p").contains("Room description").should("be.visible");
+        cy.get("p").should("have.text", "Room description").and("be.visible");
       });
 
     // Edit description
-    cy.get(".tiptap").clear();
+    clearTiptapContent();
     cy.get(".tiptap").type("New test description");
 
     // Save description
@@ -237,29 +244,37 @@ describe("Rooms view description", function () {
     cy.get('[data-test="room-description-viewer"]')
       .should("be.visible")
       .within(() => {
-        cy.get("p").contains("New test description").should("be.visible");
+        cy.get("p")
+          .should("have.text", "New test description")
+          .and("be.visible");
       });
     cy.get('[data-test="tip-tap-editor"]').should("not.exist");
     cy.get('[data-test="room-description-edit-button"]').should("be.visible");
 
     // Check with different description with different options
     cy.get('[data-test="room-description-edit-button"]').click();
+    cy.get('[data-test="room-description-edit-button"]').should("not.exist");
 
     cy.get('[data-test="tip-tap-editor"]').should("be.visible");
 
     cy.get(".tiptap")
       .should("be.visible")
+      .and("have.focus")
       .within(() => {
-        cy.get("p").contains("New test description").should("be.visible");
+        cy.get("p")
+          .should("have.text", "New test description")
+          .and("be.visible");
       });
 
-    cy.get(".tiptap").type("{selectall}");
+    selectTiptapContent();
 
     cy.get('[data-test="tip-tap-menu-dropdown-item-button"]').should(
       "have.length",
       0,
     );
+
     cy.get('[data-test="tip-tap-text-type-dropdown"]').click();
+
     cy.get('[data-test="tip-tap-menu-dropdown-item-button"]').should(
       "have.length",
       4,
@@ -282,7 +297,6 @@ describe("Rooms view description", function () {
       .should("be.visible");
 
     cy.get('[data-test="tip-tap-menu-dropdown-item-button"]').eq(0).click();
-
     cy.get('[data-test="tip-tap-menu-dropdown-item-button"]').should(
       "have.length",
       0,
@@ -291,17 +305,32 @@ describe("Rooms view description", function () {
     // Check that description was changed correctly
     cy.get(".tiptap")
       .should("be.visible")
+      .should("have.focus")
       .within(() => {
-        cy.get("h1").contains("New test description").should("be.visible");
+        cy.get("h1")
+          .should("have.text", "New test description")
+          .and("be.visible");
         cy.get("p").should("not.exist");
       });
 
-    cy.get(".tiptap").type("{selectall}");
+    cy.window().should((win) => {
+      const selection = win.getSelection();
+      const selectedText = selection.toString();
+      expect(selectedText).to.eq("New test description");
+    });
+
     cy.get('[data-test="tip-tap-bold-button"]')
       .should("be.visible")
       .and("have.class", "p-button-secondary")
       .and("have.attr", "data-p-severity", "secondary")
       .click();
+
+    cy.window().should((win) => {
+      const selection = win.getSelection();
+      const selectedText = selection.toString();
+      expect(selectedText).to.eq("New test description");
+    });
+
     cy.get('[data-test="tip-tap-bold-button"]')
       .should("be.visible")
       .and("have.class", "p-button-primary")
@@ -310,11 +339,12 @@ describe("Rooms view description", function () {
     // Check that description was changed correctly
     cy.get(".tiptap")
       .should("be.visible")
+      .and("have.focus")
       .within(() => {
         cy.get("h1")
           .find("strong")
-          .contains("New test description")
-          .should("be.visible");
+          .should("have.text", "New test description")
+          .and("be.visible");
         cy.get("p").should("not.exist");
       });
 
@@ -350,8 +380,8 @@ describe("Rooms view description", function () {
       .within(() => {
         cy.get("h1")
           .find("strong")
-          .contains("New test description")
-          .should("be.visible");
+          .should("have.text", "New test description")
+          .and("be.visible");
         cy.get("p").should("not.exist");
       });
     cy.get('[data-test="tip-tap-editor"]').should("not.exist");
@@ -457,7 +487,7 @@ describe("Rooms view description", function () {
     cy.get('[data-test="room-description-viewer"]')
       .should("be.visible")
       .within(() => {
-        cy.get("p").contains("Room description").should("be.visible");
+        cy.get("p").should("have.text", "Room description").and("be.visible");
       });
 
     cy.contains("rooms.description.missing").should("not.exist");
@@ -470,7 +500,7 @@ describe("Rooms view description", function () {
     cy.get(".tiptap")
       .should("be.visible")
       .within(() => {
-        cy.get("p").contains("Room description").should("be.visible");
+        cy.get("p").should("have.text", "Room description").and("be.visible");
       });
 
     cy.get('[data-test="room-description-viewer"]').should("not.exist");
@@ -495,7 +525,7 @@ describe("Rooms view description", function () {
     cy.get(".tiptap")
       .should("be.visible")
       .within(() => {
-        cy.get("p").contains("Room description").should("be.visible");
+        cy.get("p").should("have.text", "Room description").and("be.visible");
       });
 
     cy.get('[data-test="room-description-viewer"]').should("not.exist");
@@ -507,7 +537,9 @@ describe("Rooms view description", function () {
     cy.get('[data-test="room-description-viewer"]')
       .should("be.visible")
       .within(() => {
-        cy.get("p").contains("Changed room description").should("be.visible");
+        cy.get("p")
+          .should("have.text", "Changed room description")
+          .and("be.visible");
       });
     cy.get('[data-test="tip-tap-editor"]').should("not.exist");
   });
@@ -532,7 +564,7 @@ describe("Rooms view description", function () {
     cy.get('[data-test="room-description-viewer"]')
       .should("be.visible")
       .within(() => {
-        cy.get("a").contains("Test Link").click();
+        cy.get("a").should("have.text", "Test Link").click();
       });
 
     // Check that modal is shown
@@ -558,7 +590,7 @@ describe("Rooms view description", function () {
     cy.get('[data-test="room-description-viewer"]')
       .should("be.visible")
       .within(() => {
-        cy.get("a").contains("Test Link").click();
+        cy.get("a").should("have.text", "Test Link").click();
       });
 
     cy.window().then((win) => {