+++ tests/Frontend/e2e/RoomsViewMembersMemberActions.cy.js
@@ -1,5 +1,4 @@
 import { interceptIndefinitely } from "../support/utils/interceptIndefinitely.js";
-
 describe("Rooms view members member actions", function () {
   beforeEach(function () {
     cy.init();
@@ -32,10 +31,21 @@ describe("Rooms view members member actions", function () {
       statusCode: 204,
     }).as("userSearchRequest");
 
-    cy.get(".multiselect__content").should("not.be.visible");
+    // Check autofocus
+    cy.get(".multiselect__content").should("be.visible");
+
+    // Check prompt
+    cy.get('[data-test="select-user-dropdown"]')
+      .should("include.text", "rooms.members.modals.add.no_options")
+      .and("include.text", "rooms.members.modals.add.no_result");
 
     cy.get('[data-test="select-user-dropdown"]')
-      .should("include.text", "rooms.members.modals.add.placeholder")
+      .find("input")
+      .should(
+        "have.attr",
+        "placeholder",
+        "rooms.members.modals.add.placeholder",
+      )
       .click();
 
     cy.get('[data-test="select-user-dropdown"]').find("input").type("L");