+++ CHANGELOG.md
@@ -25,6 +25,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reload room type list on form validation error in new room dialog ([#1523], [#1616])
 - Room overview buttons not disabled during loading ([#1595])
 - Form errors not always cleared correctly ([#1679], [#1702])
+- Unnecessary loading of server pools in the room types read-only view ([#1721], [#1724])
+- Unnecessary loading of server in the server pools read-only view ([#1721], [#1724])
+- Missing permissions inheritance for user index view ([#1724])
+- Wrong permission check when showing create new server pool button ([#1724])
+- Styling issue in the role selection dropdown on the user index view ([#1724])
+- Broken permission checks when loading the user view/edit page in the admin UI ([#1724])
+
+### Removed
+
+- Unnecessary permission inheritance for the roomTypes.view and serverPools.view permissions ([#1721], [#1724])
 
 ## [v4.1.2] - 2024-11-22
 
@@ -300,6 +310,8 @@ You can find the changelog for older versions there [here](https://github.com/TH
 [#1679]: https://github.com/THM-Health/PILOS/issues/1679
 [#1702]: https://github.com/THM-Health/PILOS/pull/1702
 [#1704]: https://github.com/THM-Health/PILOS/pull/1704
+[#1721]: https://github.com/THM-Health/PILOS/issues/1721
+[#1724]: https://github.com/THM-Health/PILOS/pull/1724
 [unreleased]: https://github.com/THM-Health/PILOS/compare/v4.1.2...develop
 [v3.0.0]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.0
 [v3.0.1]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.1